requirements.txt
requests==2.32.3
pandas==2.3.0
psycopg2-binary==2.9.11